Flytt domeneklasser for registerhistorikk til egen fil
